refactor: Common 컴포넌트 디렉토리 구조 개선 및 네이밍 변경
