+++ apps/client/src/app/cart/page.tsx
@@ -4,8 +4,8 @@ import { useRouter } from 'next/navigation';
 import { useMemo } from 'react';
 import { toast } from 'sonner';
 
-import BottomNotice from '@/components/common/BottomNotice';
 import BottomFixedButton from '@/components/common/Button/BottomFixed';
+import BottomNotification from '@/components/common/Notifications/BottomNotification';
 import CartItem from '@/components/features/cart/CartItem';
 import CartSummary from '@/components/features/cart/CartSummary';
 import EmptyCart from '@/components/features/cart/EmptyCart';
@@ -132,7 +132,7 @@ export default function CartPage() {
                             </div>
                         </>
                     )}
-                    <BottomNotice type="cart" />
+                    <BottomNotification type="cart" />
                     <BottomFixedButton
                         text="구매하기"
                         isDisabled={false}
