+++ apps/client/src/app/filter/layout.tsx
@@ -1,4 +1,4 @@
-import SuspenseWrapper from '@/components/common/SuspenseWrapper';
+import SuspenseWrapper from '@/components/common/Wrappers/SuspenseWrapper';
 
 export default function Layout({ children }: { children: React.ReactNode }) {
     return <SuspenseWrapper>{children}</SuspenseWrapper>;
