+++ apps/client/src/app/oauth/kakao/layout.tsx
@@ -1,4 +1,4 @@
-import SuspenseWrapper from '@/components/common/SuspenseWrapper';
+import SuspenseWrapper from '@/components/common/Wrappers/SuspenseWrapper';
 
 export default function KakaoRediretLayout({
     children,
