+++ apps/client/src/app/order/complete/layout.tsx
@@ -1,4 +1,4 @@
-import SuspenseWrapper from '@/components/common/SuspenseWrapper';
+import SuspenseWrapper from '@/components/common/Wrappers/SuspenseWrapper';
 
 export default function OrderCompleteLayout({
     children,
