+++ apps/client/src/components/features/cart/CartItem.tsx
@@ -2,8 +2,8 @@
 
 import Image from 'next/image';
 
+import EditDeleteButton from '@/components/common/Button/EditDelete';
 import CheckBox from '@/components/common/CheckBox';
-import EditDeleteButton from '@/components/common/EditDeleteButton';
 import { useSelectedProductsStore } from '@/hooks/stores/useSelectedProductsStore';
 import { ObjectCommonInfo } from '@/schemas/object/object';
 import { formatPrice } from '@/utils/formatPrice';
